Simulated Annealing for TSP
A report focussing on deploying simulated annealing to find an optimal path connecting 30 US cities in the Travelling Salesperson Problem. The results of which yielded a shortest path connecting 30 US cities of 2176 units. This was then be scaled and benchmarked against higher order problems including Eli51 and Pr76.

Additional features of the report: 
 - Regional Cost Functions
 - Stochastic Sampling
 - Criteria Inversion