Extension to McKenna Tuama's Tacoma Narrows model.
Contained within this report is an investigation into a cornerstone of modern science. Using computational methods including forward integration techniques, the Tacoma Narrow's Bridge collapse of 1940 will be explained from a torsional viewpoint. By doing this, we will explore the effect of wind, resonance and chaos theory.

Additional features of the report:
 - 1st Order / 2nd Order Integrators.
 - Resonance Discussion.
 - Chaotic Interpretation.
 - Electromagnetic Damping Solutions.